Aggressively apply const to the rest of the code.
To me, this gives a lot more explanation as to what is going on.

Note on pointers:
int*: pointer to int
int const*: pointer to const int, this is equal to const int *
int * const: const pointer to int
int const * const: const pointer to const int., this is equal to const int * const.