Pass the sequence stack as * instead of *&.
Here, again, we should think about the semantics.
Passing as *& is only needed if we want to modify the actual pointer that we are passing, outside of the scope of the function we are calling. In the code, we were only changing the object that we are pointing to, so no need to pass as *&; removing this because *& deteriorates readability.

To illustrate the difference:
f1(int *& p1) {
  p1 = new int;
}

f2(int* p2) {
  p2 = new int;
}

int main()
{
  int* p1 = NULL;
  int* p2 = NULL;

  f1(p1); // after this call, p1 != NULL
  f2(p2); // after this call, still p2 == NULL

  return 0;
}

(cherry picked from commit aef2855a081ccbe765cdeafefa1a5fafdd07564a)