Fix bug in the handling of Replace.
The sub parameter was passed in as a pointer, and modified in the do_proof_replace and do_proof_place_replace methods. It is expected (but not yet confirmed) this issue was already present in the original implementation. Since sub is modified, the modified version can be used in subsequent proofs of other subgoals, that should not be calculated under the replacement.

Added a testcase that shows this fails.