Implement computation to establish side condition in forall with placeholder.
The existing code was, in fact, correct. We document it, and reimplement it so that the code is more self-explanatory. Note that it is slightly less efficient due to additional copying, but first correctness, and let's handle performance later.

Now, I still need to look at the rules for relativized forall, before we can put the final nail in the coffin of understanding and documenting the implementation of the proof rules!