Make bound a proper class.
This is used to present the fields in a DBM. In release mode, the compiler does a good job of inlining everything.
The memory used by the class is equal to sizeof(clock_constraint_t).

The advantage is that the interface becomes much cleaner, and in the rest of the code bound_t and clock_constraint_t cannot be confused.